Extract common string prim code
